requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-huggingface
chromadb
sentence-transformers
python-dotenv
pypdf
langchain-text-splitters
pysqlite3-binary; sys_platform == 'linux'
